(Thrift/Streaming) Introduce "sink" keyword in Thrift IDL. Add generated code for client side.
Summary:
This diff introduced the keyword "sink", which will present a client to server's stream followed by a server to client final response or final response error. Typically this feature can be used for things like client to server file upload.

The word sink, was used as an abstracted object that client can get from RPC to server. This object, from client perspective, represent server as a receiver of a flow of payloads that client could sent to server, hence we use the word "sink" to represent this relationship.

Reviewed By: jmswen

Differential Revision: D15908574

fbshipit-source-id: d68af3eaf45e2790728b611b30879dcffc0ea77a