Allow Enum as Frozen HashMap or HashSet key
Summary:
This is a long existing issue. When Enum is used as the HashMap or HashSet key, compiler error will happen like this:
    error: reference to type 'const int' could not bind to an lvalue of type 'const typename SelfKey<MetadataSource>::KeyType'

The reason is `EnumLayout` hasn't implement the `hash` function, so the `hash` function of `PackedIntegerLayout` will be used, and it only takes `const int` as key type, not the enum.

The fix is simple, just implement hash function for `EnumLayout`.

Reviewed By: yfeldblum, ddrcoder

Differential Revision: D17223958

fbshipit-source-id: 1ff3c3518a06eb6fa3652bfb4945210c1e658351