Allow Thrift randomizer to create unions with defaults
Summary:
In Thrift IDL it's possible to define a union that has default value set. This will mean the init method has a kwarg with a non-none value. The thrift randomizer will only set the arg for generated values. If the field chosen for the default, the init method will raise an exception.

This diff changes that to set all other values to none explicitly.

Reviewed By: yfeldblum, jermenkoo

Differential Revision: D17364382

fbshipit-source-id: d26e84bcd61489f63edb744e222c98fbd61d4897