Deterministic 'debugString' by sorting keys
Summary:
Diffing debug logs would be a lot more effective if it weren't randomized by hash ordering.

This diff introduces an indirect sort for `DebugProtocol` when printing associative containers. It's gated by a static bool, so it should have no impact on any non-debug behavior, though it may introduce the need for a comparator for the occasional map key.

Reviewed By: yfeldblum

Differential Revision: D17367636

fbshipit-source-id: 2e10d2fa708df76f6123d9af05f267231c9f014b