support 255+ fields on Python (>=3 && <=)3.6
Summary:
Python (>=3 <=)3.6 doesn't support function definition that takes more than 255 args.

For thrift struct with 255+ fields, the generated initializer will not compile because of that.

In order to support 255+ fields, this diff makes the struct initializer take a single `kwargs` and extract the fields within the function body, (when necessary).

This is a kinda ugly workaround and I'm really not sure whether it's the right thing to do. But it seems to work.

Reviewed By: nanshu

Differential Revision: D17489396

fbshipit-source-id: b30898140c2e334ad8cd5a1087af7019e8f7c1cb