await on get_future() when .serve is canceled
Summary:
If ThriftServer.serve() is canceled before the c++ server starts, calling .stop() on the server may not prevent it from continuing to start, leaving a thread running the thrift server after cancel.

To prevent that situation, we try to await on the get_address() since it finishes once the server serve method binds to port and seems to be in a good enough state to be stopped.

If the serve() method is canceled twice then we go back to the previous behavior of saying we are canceled and calling .stop() on ourselves but its possible python refuses to exit because the server is still running in a thread.

manually calling .stop on a thrift server from python immediately after tasking .serve() is still racy,  we should probably make stop() a coroutine and have it also await on the address_future

Reviewed By: yfeldblum

Differential Revision: D17411674

fbshipit-source-id: 6bc077001873515d20c4b86c7b5d053d00e517a1