provide a get_context() using py3.7 contextvars
Summary:
Support a more typing friendly get_context() on python 3.7 using context vars to capture the RequestContext object for the life of the thrift handler method task.

On python < 3.7 get_context() still exists but always raises a RuntimeError, this makes things a little simpler from a typing perspective. and more informative than an ImportError when trying to import get_context()

Reviewed By: nanshu

Differential Revision: D17645341

fbshipit-source-id: 455bf852cbe944239127ad3321132b392fa0f3ca