decouple shared RW state inside of CompactProtocol, FramedTransport
Summary:
This change a requirement to the thrift stack to be able to implement proper go thrift server request pipelining.

simultaneous reading and writing from different threads on a (conn, socket, stream) (thrift transport) is not a fundamental limitation of anything except the implementation of thrift transports and protocols. I found the CompactProtocol and FramedTransport both to have this behavior, where reads and writes shared some of the same state, and so reads and writes cannot happen at the same time. This decouples the state for CompactProtocol and for FramedTransport, which I believe is the common case for anything running Go thrift servers.

the parallel tests quickly discovered the limitations here, so after fixing I ran it through those tests with no problem anymore.

Reviewed By: chinmay87, lclarkmichalek

Differential Revision: D17684558

fbshipit-source-id: 195bfe6c26740cd885aa98ae1f61b71ee9bce52e