(Thrift)(Streaming)(RFC) Use Rocket[Stream|Sink]Callback* instead of [Stream|Sink]Callback* for RocketServerHandler API
Summary: some requestContext might need to preserve in the per stream/sink callback object, this API change is a way to open a chance for doing that. But there might be other way, like adding an new interface like setRequestContext on lowlevel API, but this might pollute the client side as this setRequestContext() interface will only make sense for server.

Reviewed By: andriigrynenko

Differential Revision: D17671062

fbshipit-source-id: 590571fccd3b4146e25bdd8446062a792b43f3dd