Do not parse ::1
Summary:
After working on T55843480, I noticed that a non negligible part of cpu time of
get_client() is spent parsing the parameter host, which is ::1 by default.
The return value of ipaddress.ip_address() is not used so, in the default case, parsing can be skipped.

pyperf data https://fburl.com/scuba/1yqjasbl

Reviewed By: heavenzjq

Differential Revision: D17982419

fbshipit-source-id: f5043b19ba32d84bb192bc0a2bbeb9dad41131ff