Add Instrumentation for Thrift requests
Summary:
To observe the state of thrift servers, one thing can be done is to record all the active requests in memory, so that we can both observe the states on the fly, and in coredumps.

For the active requests, we can maintain a linked list of request stub object in the class `ActiveRequestsRegistry`. The objects are inserted/removed in the list during construction/destruction. `ActiveRequestsRegistry` is owned by each of the `Cpp2Worker`.

Currently `Cpp2Request` and all the rocket request classes are supported for instrumentation. Support for `ThriftRequest` and RSocket requests are currently dropped.

Reviewed By: andriigrynenko

Differential Revision: D17530740

fbshipit-source-id: 3639922f2f7d018dfb6e34395de1e831ee885d45