Java: Dynamically load slf4j, so that Android doesn't need the dependency
Summary:
This diff introduce a `Logger` abstraction which hide the actual logger
implementation. If slf4j is in the classpath it will use it, otherwise it uses
`System.err.println`.

The prupose of this abstraction is to be able to cut the dependency on slf4j
which is currently not desired in Android.

Reviewed By: vitaut

Differential Revision: D18092663

fbshipit-source-id: 929a6df69de70a1e70e91aa0fb40b598dbddbbab