adding a Visitor interface to generated thrift union code
Summary:
Currently, java code generated for thrift union is not very easy to use. Customer need to call `isSetX()` method and then call `getX()` method to get the underlying type. This can be error prone, since there is no way to guarantee that all the cases of `isSetX()` is called. Also, such nested if statement is extremely tedious to write even for simple cases.

In this diff, we introduce a `Visitor` interface and `accept(...)` method to thrift union to make it easier to use.
#forcetdhashing

Reviewed By: stevegury

Differential Revision: D18282101

fbshipit-source-id: 66d00432d29662500e964bc32b0678c20005dcd1