Rename --gen rust:add_serde_derives to rust:serde
Summary:
The "derives" part of the original name is not meaningful because whether those impls are derived or generated by fbthrift is not visible in the API and is not something a caller should need to care about.

The new compact name is consistent with how thrift options in other language generators are named.

```
--gen cpp2:json,reflection
--gen py3:json
```

Reviewed By: jsgf

Differential Revision: D18310008

fbshipit-source-id: b345e6f84cf416f39322651e24b9b992f057444f