Escape struct and field names more consistently
Summary:
This diff updates all uses of a type's name outside of quoted strings in the generated code to mangle names that are Rust keywords. For now they are mangled by suffixing with `_` but in T45205296 we intend to change this to raw identifiers instead to stay as close as possible to what's written in the source.

For example, the Display impl for an enum which used to contain:

```
{{enum:name}}::{{enumValue:name}} => "{{enumValue:name}}"
```

now contains:

```
{{enum:rust_name}}::{{enumValue:rust_name}} => "{{enumValue:name}}"
```

Notice that the string literal's contents do not change i.e. this does not affect things like Display or FromStr or json keys.

Reviewed By: Imxset21

Differential Revision: D18320150

fbshipit-source-id: 945ab63aae999ea960d47e1ea781677f56350280