introduce a test case demonstrating the need for not erase trailing underscore char when mangling java names
Summary:
In this diff, we introduce a new test case demonstrating when it is useful to preserve trailing underscore character when mangling java names.

In the longer run, we should in theory stop mangling java names, and use language specific annotations to handle such cases. but for now, let's accept this temporary workaround.

Reviewed By: stevegury

Differential Revision: D18358505

fbshipit-source-id: 2eca87478c175e3e4221d5a6538c6cd5607a476c