Fixes for hack generation when enums are used as exception codes
Summary:
There is a change coming to Hack that improves fixme application on declarations. It prevents declaration fixmes from hiding expression level problems in a variety of cases, and changes the error code to be more accurate in certain other cases.

This diff fixes two such side effects in thrift generation, both having to do with using enums as exception codes. The first change duplicates certain fixmes on the code declaration itself. Once the new version of hack goes out, we'll look to de-dupe the fixmes, since 4341 (bad property override) is the correct one there. Also, I added a slight improvement here since this is only actually an error with non-transparent enums.

The second is on assignment to the enum property from mixed. The generated `(int)` cast works, but causes typechecker errors at the callsite. That error was hidden by the declaration fixme on the code property itself in old versions of hack. The in-flight Hack fixes will properly identify that error.

Reviewed By: DavidSnider

Differential Revision: D18357522

fbshipit-source-id: 707947707336cd96e1ff148cf511e10078138ace