Decouple Rust runtime from Tokio
Summary:
The client half of the Rust Thrift runtime was tightly tied to tokio-service. Tokio-service is deprecated and won't ever move off of tokio 0.1 and futures 0.1, so breaking this coupling is a prerequisite for upgrading the whole stack to standard library futures.

This diff eliminates all Tokio dependency from fbthrift, including tokio-codec, tokio-io, tokio-proto, and tokio-service.

Reviewed By: Imxset21

Differential Revision: D18411564

fbshipit-source-id: 554a6a1a821fbab57cdd6f5bc9f81237a20a77b8