(Thrift)(Streaming) Using TApplicationError to notify server when sink() is not called
Summary:
In bidirectional streaming, most cases cancel can be treated as a special case of error, so to make things less confusing we prefer to use error as much as possible.

For the context: we are in the direction to remove onStreamCancel API, this diff prevent cancel frame appears on the wire.

Reviewed By: iahs

Differential Revision: D17817606

fbshipit-source-id: 57f5eeb4dcae19861093b83f6818cedfdc82d09a