Make Thrift java-swift codegen works with large struct
Summary:
Pretending this diff, java-swift codegen would generate Java classes for each Struct that were created using the constructor. This is ok for the creation of “small” structs and  will not change with this diff. However, Java methods has a limitation of “256 total lengths” for class method parameters (see details below).  Basically, that limitation is 127 arguments because double/long contributes two units to the length and a parameter of any other type contributes one unit, including “this”.
With this diff Structs with more than 127 elements will be treated as a “big Struct” and created using a builder method instead of using the constructor.

******
Java 11 Documents: https://docs.oracle.com/javase/specs/jvms/se11/html/jvms-4.html
Ch. 4.3.3. Method Descriptors

A method descriptor is valid only if it represents method parameters with a total length of 255 or less, where that length includes the contribution for this in the case of instance or interface method invocations. The total length is calculated by summing the contributions of the individual parameters, where a parameter of type long or double*** contributes two units to the length and a parameter of any other type contributes one unit.

Additional good reading: When Arguments Get out of Hand
https://www.javaspecialists.eu/archive/Issue059.html

#forcetdhashing

Reviewed By: stevegury

Differential Revision: D18377451

fbshipit-source-id: 2c9c57961058cb1995113e05e67484a708931e6d