Bug fix to enable re-generate on thrift file change
The current cmake macros does not re-generate code when the thrift file is changed. This is because the cmake_custom_command only lists thrift1 as dependency. By adding the thrift file to the dependencies list, we can force cmake re-generates code when the file is changed.

Tested on my custom code.