Make the constructor of `DebugStub` take request payload.
Summary:
Motivation: It's more useful if `DebugStub` can track the request payload to help further diagnose request data. To keep track of request payload, we need to keep a clone of the IOBuf in the stub.

Some structural changes need to be made to plumb the clone of payload data into `DebugStub`. This diff only implements part of the work which simply clones the buffers and pass them to the debug stub constructor. Inside `DebugStub`, we simply drop the clone for now and left the book-keeping in following diffs.

Differential Revision: D18263266

fbshipit-source-id: fa1415d6b8f95e9fd2979364912d15f486a139dd