python TSimpleJSONProtocol: add custom skip method
Summary:
As described in T52898162, when reading a thrift struct using TSimpleJsonProtocol, if an unrecognized field contains a struct or a list, it would crash. The reason is that when the generated code sees an unrecognized field, it calls protocol.skip() and the generic version of this method (defined in TProtocol https://fburl.com/diffusion/9lua3ejv) does not work for json.

I fix this issue by creating a custom skip method for the json protocol. This is the solution the C++ implementation uses (defined here: https://fburl.com/diffusion/551h4isq)

The method I wrote is similar to the C++ one, but note that readJSONxxx methods in Python call `context.read` / `ensureAndReadContext` whereas they don't in C++.

Reviewed By: yfeldblum

Differential Revision: D18414895

fbshipit-source-id: 659acb40328b8933dd86bd7aff016d4d7be1212d