Bug fix to enable re-generate on thrift file change (#368)
Summary:
The current cmake macros does not re-generate code when the thrift file is changed. This is because the cmake_custom_command only lists thrift1 as dependency. By adding the thrift file to the dependencies list, we can force cmake re-generates code when the file is changed.

Tested on my custom code.
Pull Request resolved: https://github.com/facebook/fbthrift/pull/368

Reviewed By: vitaut

Differential Revision: D18502590

Pulled By: yfeldblum

fbshipit-source-id: 25ed8044cf99babb7578d2c8408d7b8354d8222c