Break ServerStreamFactory in two
Summary: Some of the arguments to this functor come from the generated code and others from the transport. The most natural way to pass the gencode arguments down to the transport is by partially applying them to the factory and passing that down. Here the partial application is implemented by having the SemiFactory store its arguments in the Factory lambda, which lets stream implementations take advantage of the fact that this is running on the server thread to perform some computation.

Reviewed By: andriigrynenko

Differential Revision: D18495075

fbshipit-source-id: 47e4375b58908f0488fb50d11a5acea7b005a157