Replace folly::Optional with OptionalFieldView in the API
Summary:
Replace `folly::Optional` with `OptionalFieldView` which is a new class providing `std::optional`-like API. This improves consistency between Thrift and Frozen APIs and simplifies migration from one to another. `folly::Optional` methods that are not present in `std::optional` are not provided except for three cases that are in use (`hasValue` and two comparison operators) and marked as deprecated instead.

`folly::Optional` is still used in the implementation.

Reviewed By: iahs

Differential Revision: D18534654

fbshipit-source-id: 86fefc441f7e170805e513b0c21462b0d76c5c26