remove cross-reference between py and cpp2 service interface objects
Summary: The service wrapper object is created and owned by the service interface Cython class, and it increases ref count on the python object when it was constructed, which makes the py interface unable to  destroyed. As the ThriftServer will hold the py interface object (i.e. the "handler"), it should be safe not to bump the ref count when the cpp object reference the py interface.

Reviewed By: fried

Differential Revision: D18583679

fbshipit-source-id: 7068fdaf6395c9085a7aca2de4c9c6e3beadea6e