RFC: Implement Payload Tracking in ActiveRequestsRegistry.
Summary:
Implement a general tracking mechanism for request payloads.

Each DebugStub object will keep an optional IOBuf member to its corresponding request object. To control over the total memory used by such tracking, two options, namely `maxDebugPayloadMemoryPerRequest` and `maxDebugPayloadMemoryPerWorker` are added as server attributes to give an upper bound of used memory in such cases.

Payloads are tracked by a separate intrusive list to DebugStub, and each time the memory limit is reached, payloads will be evicted based on ActiveRequestRegistry's policy. Currently the only and the default eviction policy is FIFO, which can be made to be customizable in the future work.

Differential Revision: D18335360

fbshipit-source-id: 68609b0e0e7322737618feddfefe61c206d05af9