fix thrift install location printed by `find_package(FBThrift)`
Summary:
Update the `FBThriftConfig.cmake` file to always print the correct location
where thrift was found.

Previously we could incorrectly print the location of `python-six` instead if
the fbthrift-py component was installed: due to the way CMake variable scoping
works the `find_dependency(python-six)` call would unfortunately replace the
`PACKAGE_PREFIX_DIR` variable with the prefix of `python-six`.  Therefore the
log message below would print its location.

This changes the code to store a copy of the `PACKAGE_PREFIX_DIR` variable,
and use the stored copy when logging where we found FBThrift.

Reviewed By: fanzeyi

Differential Revision: D18774510

fbshipit-source-id: 1b3c4ca0848226a5e83a6bed58c685df625f55c6