custom processing of connections based on setup frame.
Summary:
Rocket transport supports custom setup frame to be sent during connection
initiation. We take advantage of this functionality to allow intercepting
and redirecting such requests handling to custom processor.
The intended use case of this functionality is to allow admin commands, that
may not normally be part of the default thrift interface, to be accessible on
the same port that thrift server is listenting on.

Reviewed By: andriigrynenko

Differential Revision: D18596976

fbshipit-source-id: d833fc49b339084d42b961fa54bcb27945a7ad8f