Remove libthriftcpp2_shared
Summary:
There is really no need for this at all, for python3 thrift bindings to work we just build the normal `libthriftcpp2.so` and link against that everywhere. We just push the requirement that you have to build thrift with shared libraries which is natural in this case.

This not just simplified things but also fixes some fundamental linking problem where some components link against `libthriftcpp2.so` and others against `libthriftcpp2_shared.so` and this is where gflags clash in python and everything breaks.

This also breaks the installation into two steps, bdist_wheel (to offer a python wheel package) and the other steps really installs. It turned out that both cannot happen at the same time!

Reviewed By: yfeldblum

Differential Revision: D18800133

fbshipit-source-id: 70762ae5d03e69ee114cdfba352cb5b6dbbddbf4