cargo_from_buck: introduce CratesIoLockingPolicy to replace * dependencies
Summary:
In Facebook * dependencies are not a problem. The build system uses vendored crates in --offline mode, so the versions are limited to what is vendored inside tp2 crates.

Out of facebook the * dependencies can easily break a build, because authors of the dependencies might introduce API changes. That is why the * dependencies have to be replaced for some projects.

Reviewed By: StanislavGlebik

Differential Revision: D18830205

fbshipit-source-id: cd4a277e838149110985cb51173dd5e22732e185