Remove setFieldValue from TBase
Summary:
The method `setFieldValue` is designed to be a generic setter, which can set a
field based on its `field_id` and an `Object`. It is unsafe since the object is
untyped, moreover the method is not used in our android codebase (since objects are immutable).

In order to reduce the size of the generated code for android, this diff remove this method.
I added an static helper in `TBaseHelper` in order to provide the same behavior (for backward compatibility).

__Note__: The method is still generated for non-android code, but it is not part of the `TBase` interface anymore.

Reviewed By: vitaut

Differential Revision: D18418988

fbshipit-source-id: 087e0f663f68e1caf6c389de6e990f17fc3997bc