Java: Remove one of the three `toString` methods on struct
Summary:
There are 3 `toString` methods in generated classes: the default one, one with
a boolean indicating if we want to pretty print, and one with the same boolean
and an indication of the level of indentation.
This diff remove one of the `toString` in order to save space for android
codegen.

Reviewed By: vitaut

Differential Revision: D18531023

fbshipit-source-id: f0383f718c645a6750675f812e56d697d34c241e