Java: Do not generate a `read` method for Union
Summary:
Currently, the `javadeprecated` codegen generates a `read` method which
delegate the reading to the generated `readValue` method. This refactoring
remove the `read` method and move it in the parent `TUnion`.
The definition of the `metaDataMap` has moved to `TUnion` so that protocol that
requires it (`TCompactJSON`) still works.
The goal of this diff is to reduce the code size for Android.

Reviewed By: vitaut

Differential Revision: D18654541

fbshipit-source-id: a4764bdee91eda1fb82e901c785d7e6750f0e038