Look past typedef when deciding on lazy_static
Summary:
In the case of a const whose type was specified by a typedef, the Rust code generator used to determine incorrectly that a lazy_static was not needed. Example problematic Thrift code:

```
typedef map<string, Info> InfoMap

const InfoMap INFO = {
  "...": {...},
}
```

This would get emitted as a plain `const` containing a BTreeMap. This diff fixes it to emit as a lazy_static, as it would if the const's type were specified as `map<string, Info>` directly.

Reviewed By: Imxset21

Differential Revision: D18870358

fbshipit-source-id: ba10e1769a9d8357ef8a722aeb4f8e06f08e3bcd