Make the header more generic.
Summary:
The original message header was just the sizes of substreams, encoded as 32-bit
integers. This had two problems:
- Needlessly space-inefficient for small messages
- Insufficiently large for really big messages
Additionally, there was no mechanism to support the optional streams we plan
in subsequent protocol versions (initially for skipping, but things like
metadata are reasonable as well).

We solve the first two issues by making the sizes varints, and the third by
adding a flags field at the beginning.

Reviewed By: yfeldblum, bithree

Differential Revision: D17868746

fbshipit-source-id: c83bad34347d720ed748a4fde356f0ec2ac4c980