Make RequestId generation per IO thread.
Summary: The current implementation of RequestId generation places an update to a globally shared atomic counter on the hot path of Thrift server, which could impact the performance. We switch to a local generation approach where each IO thread generate their requests id locally and attach the ID of registry to the local id to construct a final RequestId.

Differential Revision: D18869673

fbshipit-source-id: 4e93b2716e6ff8aee4bca88c192692c36217987a