fix flaky tes
Summary:
Flakiness of the test is caused by server executing requests out of order.
The second request  (`wait`) is running a busy loop. We end up in a deadlock
if server starts executing it before `sendRequest` (which never gets executed),
hence the test driver will never
see that server has executed both requests (the condition it is waiting on)
Fix is to only send `wait` to the server after `sendRequest` has started
running.

(Note: this ignores all push blocking failures!)

Reviewed By: zhxchen17

Differential Revision: D19003299

fbshipit-source-id: 6f2aca4d99bd05251140194424ea116b16b8fc4f