Java: Do not generate a `read` method for Union
Summary:
Currently, the javadeprecated codegen generates a read method which
delegate the reading to the generated readValue method. This refactoring
remove the read method and move it in the parent TUnion.

A `getMetaDataMap` abstract method has been introduced in `TUnion` so
that protocol that requires it (TCompactJSON) still works.

The goal of this diff is to reduce the code size for Android.

Reviewed By: avalonalex

Differential Revision: D18948971

fbshipit-source-id: 9a2bfc6b914632474324a9740ffdaefea737af0e