Take self by shared reference in Serialize trait
Summary:
Relevant line:

```
   pub trait Serialize<P>
   where
       P: ProtocolWriter,
   {
-      fn write(self, p: &mut P);
+      fn write(&self, p: &mut P);
   }
```

This used to be the original implementation of Serialize prior to D5945856. That diff's summary says the motivation was "This makes it more flexible as it can be implemented for `&T` as well as `T`."

This diff removes all the impls for concrete reference types like &f64 and &str and replaces them with one `impl<P, T> Serialize<P> for &T where P: ProtocolWriter, T: Serialize<P>`.

Reviewed By: bolinfest

Differential Revision: D19191352

fbshipit-source-id: 722dc82ff8f1c063074d6e8f635da8ed3b4dd297