HeaderProtocol readmessagebegin should return the header seqid for header transport
Summary:
To be more compliant with fbthrift in other languages we should be using the seqid from the header instead of the one from the message.    I'm keeping the old behavior if the client types is some kind of Framed, since it doesn't have a header to store a seqid.

This should fixed thrift-py-asyncio clients talking to rust servers.

Reviewed By: nanshu

Differential Revision: D19267574

fbshipit-source-id: 1327654c7244facd8403fb4d32fe9baafddb26e9