Enable request handlers to access headers
Summary:
Until now, the go thrift library did not have an API to access headers (both
per-message and persistent headers) from a method handler.

This diff adds the Protocol object to the context handed to handlers. It's then possible to retrieve message headers from the context with HeadersFromContext, provided the communication uses a header transport/protocol.

As an optimisation, this diff doesn't add any per-message overhead: the headers map isn't copied to the context for each request. The copy only occurs if the handler code queries the headers from the context.

Reviewed By: nemith

Differential Revision: D19126763

fbshipit-source-id: af4f73b20b73254fe2093fea6131c98bc34c5c94