Cause Cython to call PyErr_Occurred() when it creates the ServiceWrapper
Summary:
https://cython.readthedocs.io/en/latest/src/userguide/language_basics.html#error-return-values

Since the import_ call is in the constructor and we don't check that is was successful or not. The internals of  import_ will set a python exception, we just have to check for it.

This has not been an issue because the module we import_ is already loaded by Python before we ever call the ctor.  But its best to cover all the bases.

Reviewed By: nanshu

Differential Revision: D19304137

fbshipit-source-id: a542a353b721434d88db9ea77dd5c8b2bbb64eb0