Remove use of [[nodiscard]] in apache::thrift::concurrency::RWGuard
Summary:
Fixes a compile-error seen in opensource builds when built using C++14.

The [[nodiscard]] attribute is only available in C++17 or later and we can't use FOLLY_NODISCARD because it might expand to `__attribute__(warn_unused_result)` which doesn't work on class definitions.

The RWGuard class is deprecated and has an easy replacement, so it's not critical for this to have the nodiscard warning enabled.

Reviewed By: jjergus

Differential Revision: D19210414

fbshipit-source-id: 15c377206e47f818c736a4a80247630cf8f64667