Do not call ServerEventHandler after Python Finalize is called
Summary:
Python is a safe environment so its never okay to abort/segfault.  There seems to be an edge case where the ServerEventHandler is called after Python has called Py_Finalize and the run-time is essentially dead.   In 3.7 onward we have a tool to check for this condition before grabbing the gil which will terminate the thread otherwise.  see: https://docs.python.org/3/c-api/init.html#c.PyGILState_Ensure

In versions less than 3.7, we just assume we are never finalizing, since a mechanism is not easily exposed yet.

Reviewed By: nanshu

Differential Revision: D19320231

fbshipit-source-id: b9c0d9c4fde15ea2aaebec1e25b60579a5af00a8