Store completion callback in smaller function holder
Summary:
This function holder is 8 bytes for the vtable pointer + the lambda itself (plus one vtable per instantiation, but that doesn't scale with the number of stream instances) while folly::Function is 64 bytes. This also prepares us for the next packing optimization by exposing the function pointer
This implementation handles several edge cases correctly that the previous implementation did not, for which regression tests are also added.

Reviewed By: andriigrynenko

Differential Revision: D18991335

fbshipit-source-id: d05ad4bf0acac600e2ec20dc1823b208bbc63520