(Thrift) back enum PROTOCOL_TYPES with uint16_t
Summary:
The largest named constant in PROTOCOL_TYPES can be represented with only 3 bits, so the range of PROTOCOL_TYPES is 3 bits. Casting values to PROTOCOL_TYPES, which are outside of the range of PROTOCOL_TYPES, is undefined behavior.

In header protocol PROTOCOL_TYPES is a 2 bytes fields, so back the enum by uint16_t.

Reviewed By: yfeldblum

Differential Revision: D19322358

fbshipit-source-id: 427118aba8b28a77ff3cee7dfad6130430e71fc0