Add enum metadata generation for modules to compiler. [1/n]
Summary:
The goal of Thrift Metadata is to expose necessary information of a given Thrift module to users as a debugging and exploring tool for services. As the very first step of this effort, we generate APIs for Thrift enum types which is called `GeneratedEnumMetadata<MyEnum>`. The IDL for describing metadata is already specified at `thrift/lib/thrift/metadata.thrift`, so we can directly generate code to fill in the information about IDL files.

This diff only adds two place placeholder files called `_metadata.cpp` and `_metadata.h` to the building targets, and currently users can only get enum information from them. The rest of metadata steup will be implemented in the following diffs.

Differential Revision: D19349903

fbshipit-source-id: 3856c7c51ea56556ddf0eebc9e43cf73dd2511b9