Service method signatures in client should take ref-of-slice for list types
Summary:
The thrift call stubs generated for Rust (before this diff) translate the thrift type `list<T>` into the argument type `&Vec<T>`, instead of `&[T]`. This means we end up losing some generality in what we can pass to these thrift stubs in the client code.

It's my understanding that it's best practice to write functions that take slice refs instead of vector refs, and I see that there's already a separate exception hard-coded in the mustache/codegen code for replacing `&String` args with `&str`... but not `Vec<T>` with `&[T]`.

Reviewed By: dtolnay

Differential Revision: D19427352

fbshipit-source-id: 6ed70f4283492a714b378b5f3d456bda875f7419