Change preprocessSendT to take THeader& directly
Summary:
The implementation of `preprocessSendT` does not actually require the ownership semantics of passing a `shared_ptr`. Passing a reference to the pointed-to object is thus more appropriate.

This allows some upcoming changes to C++ coroutine APIs for thrift clients to avoid an extra `shared_ptr` refcount increment/decrement just to pass this parameter.

Reviewed By: yfeldblum

Differential Revision: D19205139

fbshipit-source-id: 2af8cd5d2b3a8d9991373cf4f5d8382a8d6af18b