Add metadata generation for structs and exceptions in compiler [2/n]
Summary: Generates struct and exception metadata in `_metadata.cpp`. For `ThriftStruct` and `ThriftField` objects, we just assign names and ids to them. One core issue is to generate `ThriftType` in a managable way. Currently we define a set of builder objects to help constructing `ThriftType` objects for each piece of field metadata.

Differential Revision: D19394407

fbshipit-source-id: 6f42f06d9d6713e38ec6daa5850d3ed1f54e0de3