allow overriding the class used to deserialize sets and dicts
Summary:
Sets of strings in python3 are not stable and dicts change behavior,
so using repr() will give a different order depending on the seed.
This feature allows users of readFromJson to override what class
will be user to load sets and dicts, so you can use a different
implementation that keeps order or sorts the elements of the set.

For example:

  class SortedSet(set):
    def __repr__(self):
      return "{" + ", ".join(sorted(repr(x) for x in self)) + "}"

  obj = ThriftClass()
  obj.readFromJson(data, custom_set_cls=SortedSet)

Reviewed By: yfeldblum

Differential Revision: D19323374

fbshipit-source-id: 02ed95bad40860d4dadee6adcee700b1fe319688