add a _aexit_callback for attaching a guard destruction
Summary:
Add an arbitrary at __aexit__ time callback for custom client behavior with non-standard channels that may want to destroy guards when the client is out of context.

noticed we were setting the deferred_headers to None twice, and figured I would fix that logic issue while I was near by. So we clear out the vector of shared_ptr's to eventhandlers before we exit __aenter__

Reviewed By: nanshu

Differential Revision: D19180380

fbshipit-source-id: 45101c7fa213560697c0dac7f918531204592af9